Refactor Living, and make compile
