+++ src/main/java/org/granitepowered/granite/impl/entity/living/GraniteLiving.java
@@ -23,217 +23,11 @@
 
 package org.granitepowered.granite.impl.entity.living;
 
-import static org.granitepowered.granite.utils.MinecraftUtils.unwrap;
-import static org.granitepowered.granite.utils.MinecraftUtils.wrap;
-
-import com.flowpowered.math.vector.Vector3f;
-import com.google.common.base.Optional;
-import org.apache.commons.lang3.NotImplementedException;
-import org.granitepowered.granite.impl.entity.GraniteEntity;
-import org.granitepowered.granite.mappings.Mappings;
-import org.granitepowered.granite.mc.MCDamageSource;
 import org.granitepowered.granite.mc.MCEntityLiving;
-import org.granitepowered.granite.mc.MCEntityLivingBase;
-import org.spongepowered.api.entity.Entity;
-import org.spongepowered.api.entity.living.Living;
-import org.spongepowered.api.potion.PotionEffect;
-import org.spongepowered.api.potion.PotionEffectType;
-import org.spongepowered.api.world.Location;
 
-import java.util.Collection;
-import java.util.List;
-
-public class GraniteLiving<T extends MCEntityLiving> extends GraniteEntity<T> implements Living {
-
-    public GraniteLiving(T obj) {
+public class GraniteLiving extends GraniteLivingBase<MCEntityLiving> {
+    public GraniteLiving(MCEntityLiving obj) {
         super(obj);
-    }
-
-    @Override
-    public void damage(double amount) {
-        try {
-            obj.damageEntity((MCDamageSource) Mappings.getField("DamageSource", "generic").get(null), (float) amount);
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public double getHealth() {
-        return obj.getHealth();
-    }
-
-    @Override
-    public void setHealth(double health) {
-        obj.setHealth((float) health);
-    }
-
-    @Override
-    public double getMaxHealth() {
-        return obj.getMaxHealth();
-    }
-
-    @Override
-    public void setMaxHealth(double maxHealth) {
-        // TODO: Check if possible
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public void addPotionEffect(PotionEffect potionEffect, boolean force) {
-        // TODO: Potion effects, see EntityLivingBase.activePotionsMap (xo.g)
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public void addPotionEffects(Collection<PotionEffect> potionEffects, boolean force) {
-        // TODO: Potion effects, see EntityLivingBase.activePotionsMap (xo.g)
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public void removePotionEffect(PotionEffectType potionEffectType) {
-        // TODO: Potion effects, see EntityLivingBase.activePotionsMap (xo.g)
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public boolean hasPotionEffect(PotionEffectType potionEffectType) {
-        // TODO: Potion effects, see EntityLivingBase.activePotionsMap (xo.g)
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public List<PotionEffect> getPotionEffects() {
-        // TODO: Potion effects, see EntityLivingBase.activePotionsMap (xo.g)
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public Optional<Living> getLastAttacker() {
-        MCEntityLivingBase living = obj.fieldGet$lastAttacker();
-        return Optional.fromNullable((Living) wrap(living));
-    }
-
-    @Override
-    public void setLastAttacker(Living lastAttacker) {
-        obj.fieldSet$lastAttacker((MCEntityLiving) unwrap(lastAttacker));
-    }
-
-    @Override
-    public boolean isLeashed() {
-        return obj.fieldGet$isLeashed();
-    }
-
-    @Override
-    public void setLeashed(boolean leashed) {
-        obj.fieldSet$isLeashed(leashed);
-    }
-
-    @Override
-    public Optional<Entity> getLeashHolder() {
-        // TODO: Not sure what the field "leashedToEntity" in EntityLiving refers to, leaving this unimplemented for now
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public void setLeashHolder(Entity entity) {
-        // TODO: Not sure what the field "leashedToEntity" in EntityLiving refers to, leaving this unimplemented for now
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public double getEyeHeight() {
-        return obj.getEyeHeight();
-    }
-
-    @Override
-    public Vector3f getEyeLocation() {
-        // TODO: Wait for location/position shiz
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public int getRemainingAir() {
-        return obj.getAir();
-    }
-
-    @Override
-    public void setRemainingAir(int air) {
-        obj.setAir(air);
-    }
-
-    @Override
-    public int getMaxAir() {
-        // TODO: I don't think Minecraft has a concept of maximum air
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public void setMaxAir(int air) {
-        // TODO: I don't think Minecraft has a concept of maximum air
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public double getLastDamage() {
-        return obj.fieldGet$lastDamage();
-    }
-
-    @Override
-    public void setLastDamage(double damage) {
-        obj.fieldSet$lastDamage((float) damage);
-    }
-
-    @Override
-    public int getInvulnerabilityTicks() {
-        return obj.fieldGet$hurtResistantTime();
-    }
-
-    @Override
-    public void setInvulnerabilityTicks(int ticks) {
-        obj.fieldSet$hurtResistantTime(ticks);
-    }
-
-    @Override
-    public int getMaxInvulnerabilityTicks() {
-        // Minecraft stores this as half-ticks for some reason (default is 20 = 0.5 seconds, wut?)
-        return obj.fieldGet$maxHurtResistantTime() / 2;
-    }
-
-    @Override
-    public void setMaxInvulnerabilityTicks(int ticks) {
-        obj.fieldSet$maxHurtResistantTime(ticks * 2);
-    }
-
-    @Override
-    public boolean getCanPickupItems() {
-        return obj.fieldGet$canPickUpLoot();
-    }
-
-    @Override
-    public void setCanPickupItems(boolean canPickupItems) {
-        obj.fieldSet$canPickUpLoot(canPickupItems);
-    }
-
-    @Override
-    public String getCustomName() {
-        return obj.getCustomNameTag();
-    }
-
-    @Override
-    public void setCustomName(String name) {
-        obj.setCustomNameTag(name);
-    }
-
-    @Override
-    public boolean isCustomNameVisible() {
-        return obj.getAlwaysRenderNameTag();
-    }
-
-    @Override
-    public void setCustomNameVisible(boolean visible) {
-        obj.setAlwaysRenderNameTag(visible);
     }
 
     @Override
@@ -247,42 +41,22 @@ public class GraniteLiving<T extends MCEntityLiving> extends GraniteEntity<T> im
     }
 
     @Override
-    public boolean setLocation(Location location) {
-        throw new NotImplementedException("");
+    public boolean getCanPickupItems() {
+        return obj.fieldGet$canPickUpLoot();
     }
 
     @Override
-    public Optional<Entity> getPassenger() {
-        throw new NotImplementedException("");
+    public void setCanPickupItems(boolean canPickupItems) {
+        obj.fieldSet$canPickUpLoot(canPickupItems);
     }
 
     @Override
-    public Optional<Entity> getVehicle() {
-        throw new NotImplementedException("");
+    public boolean isLeashed() {
+        return obj.fieldGet$isLeashed();
     }
 
     @Override
-    public Entity getBaseVehicle() {
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public boolean setPassenger(Entity entity) {
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public boolean setVehicle(Entity entity) {
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public boolean isRemoved() {
-        throw new NotImplementedException("");
-    }
-
-    @Override
-    public boolean isLoaded() {
-        throw new NotImplementedException("");
+    public void setLeashed(boolean leashed) {
+        obj.fieldSet$isLeashed(leashed);
     }
 }
